+++ src/input/index.js
@@ -33,7 +33,9 @@ const includeProps = [
   'required',
   'customProps',
   'customError',
-  'customLabel' ]
+  'customLabel',
+  'autoFocus'
+]
 
 const defaultTypeMap = {
   [inputTypes.STRING_TYPE]: InputString,